refactor to single variants list
